+++ src/inputs/input/input.jsx
@@ -657,6 +657,9 @@ class Input extends React.PureComponent {
             'label-top': newLabelPosition === 'top',
         });
 
+        const hasLabel = !isEmpty(label);
+        const hasIcon = isString(icon) || isObject(icon) || loading || type === 'number';
+
         const renderLabel = () => {
             if (!label) {
                 return null;
@@ -690,7 +693,12 @@ class Input extends React.PureComponent {
                 className={containerClasses}
                 style={style}
             >
-                {isRedacted && <InputScreenGuard hasLabel={!isEmpty(label)} />}
+                {isRedacted && (
+                    <InputScreenGuard
+                        hasIcon={hasIcon}
+                        hasLabel={hasLabel}
+                    />
+                )}
 
                 {newLabelPosition === 'top' && renderLabel()}
 
@@ -746,7 +754,7 @@ class Input extends React.PureComponent {
 
                 {newLabelPosition === 'bottom' && renderLabel()}
 
-                {(isString(icon) || isObject(icon) || loading || type === 'number') && (
+                {hasIcon && (
                     <div
                         className="input-actions"
                         // eslint-disable-next-line no-underscore-dangle